requirements.txt
pillow==11.0.0
requests==2.32.3
progressbar2==4.5.0
aiohttp==3.11.11